hw3_semantic_segmentation/requirements.txt
numpy
pillow
scikit-learn
tqdm
matplotlib
visdom